[Fix] Tree view: Under some circumstances (e.g. after removing all unused patterns in Song Cleanup and then removing all patterns in a second Song Cleanup), the complete upper tree view could be emptied, eventually leading to a crash (tx eightbitbubsy).
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@17576 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/View_tre.cpp
@@ -1121,7 +1121,8 @@ void CModTree::UpdateView(ModTreeDocInfo &info, UpdateHint hint)
 
 		for(size_t pat = sndFile.Patterns.Size(); pat < info.tiPatterns.size(); pat++)
 		{
-			DeleteItem(info.tiPatterns[pat]);
+			if(info.tiPatterns[pat])
+				DeleteItem(info.tiPatterns[pat]);
 		}
 		info.tiPatterns.resize(sndFile.Patterns.Size(), nullptr);
